Added world two enemies to the objects, added secodn world object and world 2 is now selectabe in the second slot(middle top) in the selection menu, fized postioning but in the wave editor and fraction time in the enemy spawn rate

+++ Assets/Scripts/Game/EnemySpawner.cs
@@ -116,7 +116,7 @@ public class EnemySpawner : MonoBehaviour {
 			else{
 				Debug.LogError("No script found for "+world.enemies[i].name);
 			}
-		}else if(s[0]=='t')timer+=int.Parse(s.Substring(1,1))/10;
+		}else if(s[0]=='t')timer+=float.Parse(s.Substring(1,1))/10f;
 			else if(s[0]=='T')timer+=int.Parse(s.Substring(1,1));
 	}
 	void Chose(string s)
